https://app.usemotion.com/web/pm/workspaces/mItBy-ORI682KWmH3G1Bu?task=6B2vnD1ai1AbrwvY32cef
Corrected exercise selector for phase components. Originally, it would look for exercises with the component and subcomponent id matching the phase component, regardless of what combination was intended, now it will only look for those with the desired combination.

An example would be that, if an exercise had two possible phase components, one being flexibility-strength and the other being core-stabilization, a phase component of flexibility stabilization could have had that exercise selected for it despite being invalid. This has been corrected.